Add files via upload
This is a desktop app for a payroll system. The GUI was designed with PyQt5. The data is stored in a CSV file. Considering SQL as the final database. Payslip is printed in a txt file. Considering PDF out to replace the txt file. There are further updates i can do to improve its interaction with the user. i would keep updating it until i have a complete payroll system.